ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0


